Fix broken RecurringDateRangeRule.php
Recurring date range rule would produce a lot of errors when using yearly mode. The way we checked if the date was within a recurring scope was kind of broken.

I have rewritten it to only care about the days and months and not use a convoluted moving scope. It passes all original tests